fix nasty bug on /0 prefixes, and withdraw works superfucially

+++ bgpparse.h
@@ -1,1 +1,1 @@
-extern int _msg_count;
+extern uint64_t _msg_count;
